Rename `remote.py` to `remote_v1.py`
