+++ milatools/utils/remote_v2.py
@@ -12,7 +12,7 @@ from typing import Any, Literal
 from paramiko import SSHConfig
 
 from milatools.cli import console
-from milatools.cli.remote import Hide
+from milatools.cli.remote_v1 import Hide
 from milatools.cli.utils import DRAC_CLUSTERS, MilatoolsUserError
 
 logger = get_logger(__name__)
